Move flags to their own file - also add decode helper funcs
