+++ music/tracked/it/util.go
@@ -34,66 +34,6 @@ const (
 	DefaultFineVolume = Volume(128)
 )
 
-// IMPMFlags is a set of flags describing various features in the IT file
-type IMPMFlags uint16
-
-const (
-	// IMPMFlagStereo :: On = Stereo, Off = Mono (panning enablement flag)
-	IMPMFlagStereo = IMPMFlags(1 << 0)
-	// IMPMFlagVol0Optimizations :: If on, no mixing occurs if the volume at mixing time is 0 (redundant v1.04+)
-	IMPMFlagVol0Optimizations = IMPMFlags(1 << 1)
-	// IMPMFlagUseInstruments :: On = Use instruments, Off = Use samples
-	IMPMFlagUseInstruments = IMPMFlags(1 << 2)
-	// IMPMFlagLinearSlides :: On = Linear slides, Off = Amiga slides
-	IMPMFlagLinearSlides = IMPMFlags(1 << 3)
-	// IMPMFlagOldEffects :: On = Old Effects, Off = IT Effects
-	IMPMFlagOldEffects = IMPMFlags(1 << 4)
-	// IMPMFlagEFGLinking :: On = Link Effect G's memory with Effect E/F
-	IMPMFlagEFGLinking = IMPMFlags(1 << 5)
-	// IMPMFlagMidiPitchController :: Use MIDI pitch controller, Pitch depth given by PitchWheelDepth
-	IMPMFlagMidiPitchController = IMPMFlags(1 << 6)
-	// IMPMFlagReqEmbedMidi :: Request embedded MIDI configuration
-	IMPMFlagReqEmbedMidi = IMPMFlags(1 << 7)
-)
-
-// IMPMSpecialFlags is a set of flags describing various special features in the IT file
-type IMPMSpecialFlags uint16
-
-const (
-	// IMPMSpecialFlagMessageAttached :: On = song message attached
-	IMPMSpecialFlagMessageAttached = IMPMSpecialFlags(1 << 0)
-	// IMPMSpecialFlagReservedBit1 :: Reserved
-	IMPMSpecialFlagReservedBit1 = IMPMSpecialFlags(1 << 1)
-	// IMPMSpecialFlagReservedBit2 :: Reserved
-	IMPMSpecialFlagReservedBit2 = IMPMSpecialFlags(1 << 2)
-	// IMPMSpecialFlagEmbedMidi :: MIDI configuration embedded
-	IMPMSpecialFlagEmbedMidi = IMPMSpecialFlags(1 << 3)
-	// IMPMSpecialFlagReservedBit4 :: Reserved
-	IMPMSpecialFlagReservedBit4 = IMPMSpecialFlags(1 << 4)
-	// IMPMSpecialFlagReservedBit5 :: Reserved
-	IMPMSpecialFlagReservedBit5 = IMPMSpecialFlags(1 << 5)
-	// IMPMSpecialFlagReservedBit6 :: Reserved
-	IMPMSpecialFlagReservedBit6 = IMPMSpecialFlags(1 << 6)
-	// IMPMSpecialFlagReservedBit7 :: Reserved
-	IMPMSpecialFlagReservedBit7 = IMPMSpecialFlags(1 << 7)
-	// IMPMSpecialFlagReservedBit8 :: Reserved
-	IMPMSpecialFlagReservedBit8 = IMPMSpecialFlags(1 << 8)
-	// IMPMSpecialFlagReservedBit9 :: Reserved
-	IMPMSpecialFlagReservedBit9 = IMPMSpecialFlags(1 << 9)
-	// IMPMSpecialFlagReservedBit10 :: Reserved
-	IMPMSpecialFlagReservedBit10 = IMPMSpecialFlags(1 << 10)
-	// IMPMSpecialFlagReservedBit11 :: Reserved
-	IMPMSpecialFlagReservedBit11 = IMPMSpecialFlags(1 << 11)
-	// IMPMSpecialFlagReservedBit12 :: Reserved
-	IMPMSpecialFlagReservedBit12 = IMPMSpecialFlags(1 << 12)
-	// IMPMSpecialFlagReservedBit13 :: Reserved
-	IMPMSpecialFlagReservedBit13 = IMPMSpecialFlags(1 << 13)
-	// IMPMSpecialFlagReservedBit14 :: Reserved
-	IMPMSpecialFlagReservedBit14 = IMPMSpecialFlags(1 << 14)
-	// IMPMSpecialFlagReservedBit15 :: Reserved
-	IMPMSpecialFlagReservedBit15 = IMPMSpecialFlags(1 << 15)
-)
-
 // PanSeparation is the panning separation value
 type PanSeparation uint8
